Added Lego SumoBot Code
The code for the Lego SumoBot was developed during a robotics competition. This code was made for a robot that was being built in the same short timeframe (approx 1.5 hours with a set of provided materials.) Its goal was to push the other robot out of the arena. The robot that used this code (also made by me) finished in second place of the 17 robots in the competition.